Refactor
* Changed joi validation name

* Changed wrapper name

* Added eslint script in package json

* Changet req.session.user!._id to req.session.user?._id

* Other small things

+++ server-express/src/users/users.service.ts
@@ -1,8 +1,12 @@
+import { ForbiddenException } from '../model/exceptions/forbidden.exception';
 import { ISession } from './users.interfaces';
 import { UsersRepository } from './users.repository';
 
 class UsersServiceClass {
-  async deleteAccount(id: string, session: ISession) {
+  async deleteAccount(id: string | undefined, session: ISession) {
+    if (!id) {
+      throw new ForbiddenException();
+    }
     session.user = null;
     return await UsersRepository.deleteAccount(id);
   }
